requirements.txt
Appium-Python-Client==4.0.0
pytest==8.1.1
pytest-html==4.0.2
selenium==4.11.2
python-dotenv==1.0.0
